requirements.txt
streamlit==1.30.0
openai==1.3.5
python-dotenv==1.0.0
